content/html/content/test/test_bug588683-4.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=588683
-->
<head>
  <title>Test for Bug 588683</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=588683">Mozilla Bug 588683</a>
<p id="display"></p>
<div id="content" style='display:none;'>
  <form id='f'>
  </form>
  <table id='t'>
    <form id='f2'>
      <tr><td><input id='i1'></td></tr>
      <tr><td><input id='i2' form='f'></td></tr>
    </form>
  </table>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 588683 **/

var table = document.getElementById('t');
var i1 = document.getElementById('i1');
var i2 = document.getElementById('i2');

is(i1.form, document.getElementById('f2'),
   "i1 form should be it's parent");
is(i2.form, document.getElementById('f'),
   "i1 form should be the form with the id in @form");

table.removeChild(document.getElementById('f2'));
is(i1, document.getElementById('i1'),
   "i1 should still be in the document");
is(i1.form, null, "i1 should not have any form owner");
is(i2.form, document.getElementById('f'),
   "i1 form should be the form with the id in @form");

</script>
</pre>
</body>
</html>
